Don't break on non-fused iterators
When the source iterator ends before the array buffer is filled, the buffer is left in an half-initialized state so that remainder() can yield the initialized part of the buffer. However, if the source iterator starts yielding items again, the initialized part of the buffer is overwritten (memory leak), but more importantly, the num_init variable is _not_ overwritten and counts too far. The Drop impl tries to deinitialize num_init amount of items, and therefore drops uninitialized memory